LocalExecutor/requirements.txt
pyautogui
pillow
webbrowser